VirtualWatchdog: Simplify the design
1. VirtualWatchdog may now be used on platforms without the hardware
watchdog peripheral.

2. Remove the interference between VirtualWatchdog objects & Watchdog
driver. Do not start the hardware Watchdog in the ctor of the first
instance of VirtualWatchdog. Remove the _ticker member that was used to
kick the hardware Watchdog. Both classes are independent and may be
used simultaneously.

3. Remove the static list of all active instances of VirtualWatchdog.
Remove the process() method that iterates over all active instances.
Add private _timeout object that calls system_reset() instead. The
kick() method postpones the reset (same as before). Every instance is
independent.

4. Replace assertions with no-ops.